feat: refactor images
