+++ src/images/svg/dbot/index.ts
@@ -3,8 +3,8 @@ export { default as BotIcon } from './bot-icon.svg'
 export { default as DbotBgMobile } from './dbot-bg-mobile.svg'
 export { default as DbotBg } from './dbot-bg.svg'
 export { default as BotBannerLogo } from './deriv-bot-banner-logo.svg'
-export { default as DtIcon } from './dt-icon.svg'
-export { default as DxIcon } from './dx-icon.svg'
+export { default as DtIcon } from '../custom/rebranding/dtrader-icon.svg'
+export { default as DxIcon } from '../custom/rebranding/deriv-x-icon.svg'
 export { default as EzIcon } from './ez-icon.svg'
-export { default as GoIcon } from './go-icon.svg'
+export { default as GoIcon } from '../custom/rebranding/deriv-go-icon.svg'
 export { default as Mt5Icon } from './mt5.svg'
