+++ src/images/svg/deriv-x/index.ts
@@ -4,7 +4,7 @@ import IconNewPromising from 'images/svg/deriv-x/new-promising.svg'
 import MT5Icon from 'images/svg/deriv-x/mt5.svg'
 import DTIcon from 'images/svg/deriv-x/dt-icosn.svg'
 import EZIcon from 'images/svg/deriv-x/ez-icon.svg'
-import DGoIcon from 'images/svg/deriv-x/go-icon.svg'
+import DGoIcon from 'images/svg/custom/rebranding/deriv-go-icon.svg'
 import DBIcon from 'images/svg/deriv-x/db-icon.svg'
 import DerivXLogo from 'images/svg/deriv-x/derivX.svg'
 import DerivXBannerLogo from 'images/svg/deriv-x/deriv-x-banner-logo.svg'
