+++ src/pages/deriv-go/_other-apps.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import mt5Icon from '../../images/svg/deriv-go/mt5.svg'
 import dtIcon from '../../images/svg/deriv-go/dt-icosn.svg'
 import ezIcon from '../../images/svg/deriv-go/ez-icon.svg'
-import dxIcon from '../../images/svg/deriv-go/dx-icon.svg'
+import dxIcon from '../../images/svg/custom/rebranding/deriv-x-icon.svg'
 import dbIcon from '../../images/svg/deriv-go/db-icon.svg'
 import GenericCarousel, { TRenderableData } from 'components/custom/carousel/_platforms-carousel'
