+++ src/pages/why-choose-us/_africa-forex-award.tsx
@@ -7,12 +7,10 @@ import { Flex, Desktop, Mobile } from 'components/containers'
 
 const query = graphql`
     query {
-        africa_award_background: file(relativePath: { eq: "choose-us/africa_award_bg.png" }) {
+        award_background: file(relativePath: { eq: "choose-us/award_bg.png" }) {
             ...fadeIn
         }
-        africa_award_background_mobile: file(
-            relativePath: { eq: "choose-us/africa_award_bg_mobile.png" }
-        ) {
+        award_background_mobile: file(relativePath: { eq: "choose-us/award_bg_mobile.png" }) {
             ...fadeIn
         }
         africa_award: file(relativePath: { eq: "choose-us/africa_award.png" }) {
